fix: renaming some files
