+++ src/main/java/com/blogify/blogapi/repository/CommentReactionRepository.java
@@ -1,9 +1,9 @@
 package com.blogify.blogapi.repository;
 
-import com.blogify.blogapi.model.CommentReact;
+import com.blogify.blogapi.repository.model.CommentReaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentReactRepository extends JpaRepository<CommentReact, Long> {
+public interface CommentReactionRepository extends JpaRepository<CommentReaction, Long> {
 }
