+++ src/main/resources/db/migration/V42_8__Create_comment_table.sql
@@ -1,4 +1,4 @@
-create table if not exists comments
+create table if not exists comment
 (
     id UUID PRIMARY KEY,
     content TEXT,
