--- src/main/resources/db/migration/V42_9__Create_comment_react_table.sql
@@ -1,6 +1,0 @@
-create table if not exists comment_reacts
-(
-    id UUID PRIMARY KEY,
-    comment_id UUID REFERENCES comments(id),
-    is_like BOOLEAN
-);
