+++ src/main/resources/db/migration/V42_9__Create_comment_reaction_table.sql
@@ -1,0 +1,6 @@
+create table if not exists comment_reaction
+(
+    id UUID PRIMARY KEY,
+    comment_id UUID REFERENCES comment(id),
+    is_like BOOLEAN
+);
